feat: separate album list as component
